verifying router (#12)
* rename keyword "exists" and fix merge artifact

* fix import errors (missing specs) for router package

* add spec for errors.New

* fix svc_spec_test.gobra

* remove unnecessary Mem predicates from specs as netip.AddrPort is struct that is passed by value

* fix errors spec

* some more spec

* increase timeout

* adapt Mem to changing struct entries of DataPlane

* add missing ghost parameter in Dataplane struct

* remove wrong atomic spec

fix some more type errors

* fix a lot more type errors

* add more spec

* add more spec

* some more spec

* more spec to reduce number of type errors

* more spec, less errors

* fix some more type errors

* add spec of Errorf to fmt

* fix addressable issues

* fixing some more issues

* fix slayers verification

* add tmp variable to fix issue with map element comparison in gobra

* manually dereference pointer to buffer for reslice operation.

* fix two more errors

* temporarily fix some more errors

* temporarily comment out some code to procede with verification

* add temporary variables to make them addressable

* remove unused predicate

* improve gobra shortcomming fixes by using outline blocks

* fix const in func error

* add spec to serialize_proxy.go

* clean up comments

* fix spec of seriallize_proxy

* cleanup

---------

+++ router/metrics.go
@@ -235,6 +235,9 @@ func classOfSize(pktSize int) sizeClass {
 
 // Returns a human-friendly representation of the given size class. Avoid bracket notation to make
 // the values possibly easier to use in monitoring queries.
+// @ requires  minSizeClass <= sc && sc < maxSizeClass
+// @ ensures   sc == old(sc)
+// @ decreases
 func (sc sizeClass) String() string {
 	low := strconv.Itoa((1 << sc) >> 1)
 	high := strconv.Itoa((1 << sc) - 1)
@@ -282,6 +285,9 @@ type outputMetrics struct {
 	OutputPacketsTotal prometheus.Counter
 }
 
+// @ requires acc(metrics)
+// @ requires acc(neighbors)
+// @ decreases
 func newInterfaceMetrics(
 	metrics *Metrics,
 	id uint16,
@@ -290,6 +296,9 @@ func newInterfaceMetrics(
 
 	ifLabels := interfaceLabels(id, localIA, neighbors)
 	m := interfaceMetrics{}
+	// @ invariant minSizeClass <= sc
+	// @ invariant sc <= maxSizeClass
+	// @ decreases maxSizeClass - sc
 	for sc := minSizeClass; sc < maxSizeClass; sc++ {
 		scLabels := prometheus.Labels{"sizeclass": sc.String()}
 		m[sc] = newTrafficMetrics(metrics, ifLabels, scLabels)
@@ -297,6 +306,10 @@ func newInterfaceMetrics(
 	return m
 }
 
+// @ trusted
+// @ requires false
+// @ requires metrics.Mem()
+// @ decreases
 func newTrafficMetrics(
 	metrics *Metrics,
 	ifLabels prometheus.Labels,
@@ -333,16 +346,20 @@ func newTrafficMetrics(
 	c.DroppedPacketsBusySlowPath =
 		metrics.DroppedPacketsTotal.MustCurryWith(ifLabels).MustCurryWith(scLabels).With(reasonMap)
 
-	c.InputBytesTotal.Add(0)
-	c.InputPacketsTotal.Add(0)
-	c.DroppedPacketsInvalid.Add(0)
-	c.DroppedPacketsBusyProcessor.Add(0)
-	c.DroppedPacketsBusyForwarder.Add(0)
-	c.DroppedPacketsBusySlowPath.Add(0)
-	c.ProcessedPackets.Add(0)
+	// (VerifiedSCION): added the cast to float64 as gobra does not handle implicit cast to float64.
+	c.InputBytesTotal.Add(float64(0))
+	c.InputPacketsTotal.Add(float64(0))
+	c.DroppedPacketsInvalid.Add(float64(0))
+	c.DroppedPacketsBusyProcessor.Add(float64(0))
+	c.DroppedPacketsBusyForwarder.Add(float64(0))
+	c.DroppedPacketsBusySlowPath.Add(float64(0))
+	c.ProcessedPackets.Add(float64(0))
 	return c
 }
 
+// @ trusted
+// @ requires false
+// @ requires metrics.Mem()
 func newOutputMetrics(
 	metrics *Metrics,
 	ifLabels prometheus.Labels,
@@ -354,12 +371,17 @@ func newOutputMetrics(
 		metrics.OutputBytesTotal.MustCurryWith(ifLabels).MustCurryWith(scLabels).With(ttLabels)
 	om.OutputPacketsTotal =
 		metrics.OutputPacketsTotal.MustCurryWith(ifLabels).MustCurryWith(scLabels).With(ttLabels)
-	om.OutputBytesTotal.Add(0)
-	om.OutputPacketsTotal.Add(0)
+
+	// (VerifiedSCION): added the cast to float64 as gobra does not handle implicit cast to float64.
+	om.OutputBytesTotal.Add(float64(0))
+	om.OutputPacketsTotal.Add(float64(0))
 	return om
 }
 
+// @ requires acc(neighbors)
+// @ decreases
 func interfaceLabels(id uint16, localIA addr.IA, neighbors map[uint16]addr.IA) prometheus.Labels {
+	// @ assume 0 <= id
 	if id == 0 {
 		return prometheus.Labels{
 			"isd_as":          localIA.String(),
